+++ lib/mix/tasks/timber/install/code_helper.ex
@@ -1,3 +1,0 @@
-defmodule Mix.Tasks.Timber.Install.CodeHelper do
-  @moduledoc false
-end