+++ test/support/test_case.ex
@@ -2,10 +2,6 @@ defmodule Timber.TestCase do
   use ExUnit.CaseTemplate, async: false
 
   setup _tags do
-    Timber.Installer.FakeFile.reset()
-    Timber.Installer.FakeHTTPClient.reset()
-    Timber.Installer.FakeIO.reset()
-    Timber.Installer.FakePath.reset()
     Timber.FakeHTTPClient.reset()
     :ok
   end